Polynomial Research
A randomized mode is added to the queue pop step when generating the root cloud (controllable via a flag and disabled by default) to see if it was possible to escape the usual bounded region. I believe that I observed this in some prior visualzation studies of the polynomial roots function, but I've still yet to verify it.